+++ lib/active_model_extras.rb
@@ -4,10 +4,10 @@ require "active_model"
 require "active_support"
 require "active_support/core_ext"
 
-require_relative "activemodel-extra/version"
-require_relative "activemodel-extra/nested_model"
+require_relative "active_model_extras/version"
+require_relative "active_model_extras/nested_model"
 require_relative "active_model/type/array"
 require_relative "active_model/validations/array_validator"
 
-module ActiveModelExtra
+module ActiveModelExtras
 end